requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements/base.in
#
asgiref==3.8.1
    # via django
certifi==2025.6.15
    # via requests
charset-normalizer==3.4.2
    # via requests
crispy-bootstrap4==2025.6
    # via -r requirements/base.in
django==4.2.23
    # via
    #   -r requirements/base.in
    #   crispy-bootstrap4
    #   django-appconf
    #   django-crispy-forms
    #   django-filter
    #   django-select2
    #   django-version-checks
django-appconf==1.1.0
    # via django-select2
django-crispy-forms==2.4
    # via
    #   -r requirements/base.in
    #   crispy-bootstrap4
django-environ==0.12.0
    # via -r requirements/base.in
django-filter==25.1
    # via -r requirements/base.in
django-select2==8.4.1
    # via -r requirements/base.in
django-version-checks==1.14.0
    # via -r requirements/base.in
idna==3.10
    # via requests
lxml==5.4.0
    # via python-docx
packaging==25.0
    # via django-version-checks
pillow==11.2.1
    # via -r requirements/base.in
pypdf2==3.0.1
    # via -r requirements/base.in
python-docx==1.1.2
    # via -r requirements/base.in
requests==2.32.4
    # via -r requirements/base.in
sqlparse==0.5.3
    # via django
typing-extensions==4.14.0
    # via
    #   asgiref
    #   python-docx
urllib3==2.4.0
    # via requests
